al fin las funciones anda
comenzando con los estilos y HAY QUE ARREGLAR EL LOGIN

+++ src/app/services/expense.service.ts
@@ -12,19 +12,21 @@ import { UserInfo } from '@firebase/auth-types';
 @Injectable()
 export class ExpenseService {
 
-  user: string
+  user: UserInfo
 
   constructor(private db: AngularFirestore,
     private afAuth: AngularFireAuth) {
-      this.afAuth.authState.subscribe(user => this.user = user.providerData[0].uid) 
-    }
 
-  ngOnInit() {
-    
+      this.afAuth.authState.subscribe(user => {
+        if (user) {
+           this.user = user.providerData[0]
+        }
+      });
+
   }
 
   getOwnExpenses() {
-    return this.db.collection(`users/${this.user}/expenses`).snapshotChanges()
+    return this.db.collection(`users/${this.user.uid}/expenses`).snapshotChanges()
       .map(actions =>
         actions.map(a => {
           let expense = a.payload.doc.data() as Expense
@@ -34,7 +36,7 @@ export class ExpenseService {
   }
 
   getOtherExpenses() {
-    return this.db.collection(`users/${this.user}/otherExpenses`).snapshotChanges()
+    return this.db.collection(`users/${this.user.uid}/otherExpenses`).snapshotChanges()
       .map(actions =>
         actions.map(a => {
           let expense = a.payload.doc.data() as Expense
@@ -43,8 +45,8 @@ export class ExpenseService {
         }))
   }
 
-  getExpense(id: string,type:string): Observable<Expense> {
-    return this.db.doc(`users/${this.user}/${type}/${id}`).snapshotChanges()
+  getExpense(id: string, type: string): Observable<Expense> {
+    return this.db.doc(`users/${this.user.uid}/${type}/${id}`).snapshotChanges()
       .map(doc => {
         var expense = doc.payload.data() as Expense
         expense.id = doc.payload.id;
@@ -52,44 +54,18 @@ export class ExpenseService {
       });
   }
 
-  getExpenseUsers(id: string,type:string): Observable<string[]> {
-    return this.db.collection(`users/${this.user}/${type}/${id}/users`).snapshotChanges()
-      .map(action => action.map(a => a.payload.doc.id));
-  }
-
   addExpense(expense: Expense, expenseUsers: string[]): Promise<any> {
     return this.db.collection(`users/${expense.creator}/expenses`).add(expense)
-      .then(doc => expenseUsers.forEach(user => doc.collection('users').doc(user).set({
-        id: user,
-        individualAmount: expense.totalAmount / expenseUsers.length
-      } as ExpenseUser))
-      )
   }
 
-  updateExpense(expense: Expense, expenseUsers: string[]): Promise<any> {
-    let expenseRef = this.db.doc(`users/${this.user}/expenses/${expense.id}`);
-    let expenseUsersRef = this.db.collection(`users/${this.user}/expenses/${expense.id}/users`);
-    return expenseRef.update(expense)
-      .then(_ => expenseUsersRef.ref.get())
-      .then(query => Promise.all(
-        query.docs.map(doc => {
-          if (!expenseUsers.find(eu => eu === doc.id)) { doc.ref.delete() }
-        })))
-      .then(_ => expenseUsers.map(user =>
-        expenseUsersRef.doc(user).set({
-          id: user,
-          individualAmount: expense.totalAmount / expenseUsers.length
-        } as ExpenseUser)))
+  updateExpense(expense: Expense): Promise<any> {
+    return this.db.doc(`users/${this.user.uid}/expenses/${expense.id}`).set(expense)
   }
 
   deleteExpense(expenseId: string) {
-    return this.deleteExpenseUsers(expenseId).then(_ =>
-      this.db.doc(`users/${this.user}/expenses/${expenseId}`).delete())
+    return this.db.doc(`users/${this.user.uid}/expenses/${expenseId}`).delete()
   }
 
-  deleteExpenseUsers(expenseId: string): Promise<any> { // TODO: hacerlo por cloud functions
-    return this.db.collection(`users/${this.user}/expenses/${expenseId}/users`).ref.get()
-      .then(query => query.docs.map(doc => doc.ref.delete()))
-  }
+
 
 }
